Move contents of sso.go into different files
