+++ pkg/auth/dependency/sso/authinfo.go
@@ -9,6 +9,38 @@ import (
 	"github.com/skygeario/skygear-server/pkg/core/errors"
 )
 
+// AuthInfo contains auth info from HandleAuthzResp
+type AuthInfo struct {
+	ProviderConfig          config.OAuthProviderConfiguration
+	ProviderRawProfile      map[string]interface{}
+	ProviderAccessTokenResp interface{}
+	ProviderUserInfo        ProviderUserInfo
+}
+
+type ProviderUserInfo struct {
+	ID    string
+	Email string
+}
+
+type OAuthAuthorizationResponse struct {
+	Code  string
+	State string
+	Scope string
+	// Nonce is required when the provider supports OpenID connect or OAuth Authorization Code Flow.
+	// The implementation is based on the suggestion in the spec.
+	// See https://openid.net/specs/openid-connect-core-1_0.html#NonceNotes
+	//
+	// The nonce is a cryptographically random string.
+	// The nonce is stored in the session cookie when auth URL is called.
+	// The nonce is hashed with SHA256.
+	// The hashed nonce is given to the OIDC provider
+	// The hashed nonce is stored in the state.
+	// The callback endpoint expect the user agent to include the nonce in the session cookie.
+	// The nonce in session cookie will be validated against the hashed nonce in the ID token.
+	// The nonce in session cookie will be validated against the hashed nonce in the state.
+	Nonce string
+}
+
 type getAuthInfoRequest struct {
 	urlPrefix      *url.URL
 	oauthConfig    *config.OAuthConfiguration
