+++ pkg/auth/dependency/sso/callback_url_test.go
@@ -6,17 +6,6 @@ import (
 	. "github.com/smartystreets/goconvey/convey"
 )
 
-func TestIsValidUXmode(t *testing.T) {
-	Convey("Test IsValidUXMode", t, func() {
-		So(IsValidUXMode(""), ShouldBeFalse)
-		So(IsValidUXMode("nonsense"), ShouldBeFalse)
-
-		So(IsValidUXMode(UXModeWebRedirect), ShouldBeTrue)
-		So(IsValidUXMode(UXModeWebPopup), ShouldBeTrue)
-		So(IsValidUXMode(UXModeMobileApp), ShouldBeTrue)
-	})
-}
-
 func TestValidateCallbackURL(t *testing.T) {
 	Convey("Test ValidateCallbackURL", t, func() {
 		f := ValidateCallbackURL
